update provider pattern to support one instance per producer/consumer
